Update data URL in locations.inc.php
The data URL inside the 'from_og' function in 'locations.inc.php' has been updated. It now points to a JSON file located in the 'pricing.mardens.com' domain instead of the previous TXT file in 'fm.mardens.com'. This change aligns with a switch to using JSON as our primary data format.