Created new 'History' class in PHP
This commit adds a new PHP file, 'history.inc.php', which introduces the History class. This class is responsible for handling history-related actions in the database such as retrievals and additions of history records based on provided parameters, enhancing data management and access. Included in this class are several public functions catering to specific needs, improving flexibility and functionality.