Refactor history view and initiate click-to-view-details functionality
This commit refactors the history view pop up in 'history.html'. Redundant CSS classes have been removed, and the structure has been optimized. The key enhancement is the initiation of click-to-view-details functionality that allows for viewing detailed history. Preparational changes for this feature such as setting up event listeners are included, improving the interactive capability of the history section.