Add user accountability to API endpoint actions
API location endpoint actions (POST, DELETE) have been updated to record who made changes. In these methods, the user who is performing the operation is taken from the request header "X-User". If the user is not found, the default "system" user is used. This change aimed to provide better traceability for the changes.