Update directory-list.js to enhance data types and functionality
The commit improves the directory-list.js by changing the return data type of getListItems from an Array to an Object. It has also included importing getHistory from history.js to fetch item history, resulting in a better-organized and more readable code. Additionally, the commit added a check for "from_filemaker" property in the buildListHTML function to ensure compatibility with a potential new data source.