Suppress warnings in location API
The code has been modified to suppress potential warnings in several parts of the location API by adding "@" before variable declarations. This change helps ensure that even if the variables are not set, their lack of setting will not disrupt the operation of the code. Now, the application can smoothly handle scenarios where user input or data might not be as expected.